+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- (24 Jun, 2020) Stackoverflow search (via google) added
 - (16 Jun, 2020) C++ search added
 - (20 Dec, 2019) Golang search added
 - (19 Dec, 2019) Minor fixes for PHP search